cv_agent_ai/pyproject.toml
[project]
name = "cv_agent_ai"
version = "0.1.0"
description = "AI-powered CV analysis and optimization tool"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]

dependencies = [
    "crewai>=0.11.0",
    "python-dotenv>=1.0.0",
    "PyGithub>=2.1.1",
    "pydantic>=2.5.0",
    "deepseek-ai>=0.1.0",  # Pour DeepSeek
]

[project.scripts]
cv_agent_ai = "cv_agent_ai.main:run"
run_crew = "cv_agent_ai.main:run"
train = "cv_agent_ai.main:train"
replay = "cv_agent_ai.main:replay"
test = "cv_agent_ai.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.hatch.build.targets.wheel]
packages = ["src/cv_agent_ai"]
